chore: moved to yarn workspaces to have a better testing suite
This is done because we want to resolve internal packages during e2e without using shim code.
