+++ jest.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+    testPathIgnorePatterns: ['<rootDir>/node_modules/'],
+    // transformIgnorePatterns: ['<rootDir>.*(node_modules)(?!.*webpack-cli.*).*$'],
+    testEnvironment: 'node',
+    collectCoverage: true,
+    coverageReporters: ['json', 'html', 'cobertura'],
+    transform: {
+        '^.+\\.(ts)?$': 'ts-jest',
+    },
+    testRegex: ['/__tests__/.*\\.(test.js|test.ts)$', '/test/.*\\.(test.js|test.ts)$'],
+    moduleFileExtensions: ['ts', 'js', 'json'],
+};
