+++ lerna.json
@@ -2,6 +2,8 @@
   "lerna": "3.20.2",
   "packages": ["packages/*"],
   "version": "independent",
+  "npmClient": "yarn",
+  "useWorkspaces": true,
   "command": {
     "version": {
       "message": "chore(release): publish %s",
