+++ packages/cli/__tests__/ZeroConfigGroup.test.js
@@ -1,4 +1,4 @@
-const ZeroConfigGroup = require('../packages/cli/lib/groups/ZeroConfigGroup');
+const ZeroConfigGroup = require('../lib/groups/ZeroConfigGroup');
 
 describe('GroupHelper', function() {
     it('should  load the dev zero config', () => {
