+++ packages/cli/__tests__/command/external.test.js
@@ -1,0 +1,9 @@
+const ExternalCommand = require('../../lib/commands/external')
+
+describe('external command', () => {
+    it('should check existence of package', () => {
+           const exists =  ExternalCommand.checkIfPackageExists('info')
+
+        expect(exists).toBeFalsy();
+    })
+})
