+++ test/info/info-output.test.js
@@ -5,7 +5,7 @@ const path = require('path');
 const { run } = require('../utils/test-utils');
 
 const runInfo = args => {
-    return run(path.resolve(__dirname, './shim'), ['info'].concat(args), false);
+    return run(path.resolve(__dirname), ['info'].concat(args), false);
 };
 
 describe('basic info usage', () => {
